move redux, add hooks, standardize hook import
