+++ src/app/problems/[pid]/EditorClient.tsx
@@ -1,5 +1,4 @@
 "use client";
-import { useAppSelector } from "@redux/hooks";
 import { submit } from "@utils/actions";
 import type { DetailedProblemInfo } from "@utils/api/getProblemInfo";
 import { clsx } from "@utils/shared";
@@ -8,6 +7,7 @@ import { useRouter } from "next/navigation";
 import { useEffect, useRef, useState } from "react";
 
 import atlanticNight from "@/config/monaco-themes/atlantic-night.json";
+import { useAppSelector } from "@/hooks";
 
 export interface EditorClientProps {
   problemData: DetailedProblemInfo;
