+++ src/components/NavigationBar/NavLinks.tsx
@@ -1,7 +1,5 @@
 "use client";
 import { Menu, RadioGroup, Transition } from "@headlessui/react";
-import { useAppDispatch, useAppSelector } from "@redux/hooks";
-import { setScheme, setUser } from "@redux/slice";
 import { logout } from "@utils/actions";
 import { clsx, getCodefunRole, getCodefunRoleTextClass } from "@utils/shared";
 import Image from "next/image";
@@ -9,6 +7,8 @@ import { usePathname, useRouter } from "next/navigation";
 import { useEffect, useState } from "react";
 
 import { ComputerIcon, MoonIcon, SunIcon, UserIcon } from "@/components/icon";
+import { useAppDispatch, useAppSelector } from "@/hooks";
+import { setScheme, setUser } from "@/providers/redux";
 
 import { ADDITIONAL_LINKS, SIGNED_IN_LINKS, SIGNED_OUT_LINKS } from "./constants";
 import { BaseNavLink, HorizontalNavLink, NAV_BUTTON_CLASS, VerticalNavLink } from "./NavLink";
