+++ src/hooks/useAppDispatch.ts
@@ -1,0 +1,5 @@
+import { useDispatch } from "react-redux";
+
+import type { AppDispatch } from "@/providers/redux";
+
+export const useAppDispatch: () => AppDispatch = useDispatch;
