+++ src/providers/redux/ReduxProvider.tsx
@@ -1,7 +1,8 @@
 "use client";
-import { store } from "@redux/store";
 import type { ReactNode } from "react";
 import { Provider } from "react-redux";
+
+import { store } from "./store";
 
 export const ReduxProvider = ({ children }: { children: ReactNode }) => (
   <Provider store={store}>{children}</Provider>
